+++ lib/game_result.rb
@@ -1,22 +1,3 @@
-class Move
-  def initialize(input)
-    @input = input
-  end
-  attr_accessor :input
-
-  def validate(arr, arr_two)
-    if arr.count(@input).positive? && !arr_two.count(@input).positive?
-      true
-    else
-      false
-    end
-  end
-
-  def apply_input(arr, ite)
-    arr[@input.to_i - 1] = ite.odd? ? 'X' : 'O'
-  end
-end
-
 class GameResult
   def equality_checker(arg1, arg2, arg3)
     return true if arg1 == arg2 && arg2 == arg3
